requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.4
sqlmodel==0.0.22
SQLAlchemy==2.0.36
python-multipart==0.0.20
httpx==0.27.2
apscheduler==3.10.4
itsdangerous==2.2.0
python-dotenv==1.0.1
